2021-2022/01_DeepLearning/mid-project/requirements.txt
pytorch
torchvision
torchaudio
cudatoolkit=10.1
matplotlib
networkx
tensorboardx
scikit-learn
scipy
numpy
tqdm
prettytable

